+++ lib/jay_api/elasticsearch/query_builder/aggregations/terms.rb
@@ -14,7 +14,7 @@ module JayAPI
         # Information about this type of aggregation can be found in:
         # https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-terms-aggregation.html
         class Terms < ::JayAPI::Elasticsearch::QueryBuilder::Aggregations::Aggregation
-          attr_reader :field, :script, :size, :order
+          attr_reader :field, :script, :size, :order, :missing
 
           # @param [String] name The name used by Elasticsearch to identify each
           #   of the aggregations.
@@ -29,9 +29,15 @@ module JayAPI
           #   aggregation. By default, the +terms+ aggregation orders terms by
           #   descending document +_count+. This can be changed by providing a
           #   custom +order+ hash.
+          # @param [String] missing The value to use for documents that are
+          #   missing a value in the given +field+ or for whom the +script+
+          #   returns +null+. When +missing+ is not given, such documents are
+          #   ignored.
           # @raise [ArgumentError] If neither a +field+ nor a +script+ are given
           #   or if both of them are given. Only one should be present.
-          def initialize(name, field: nil, script: nil, size: nil, order: nil)
+          # rubocop:disable Metrics/ParameterLists -- Constraint by Elasticsearch's design
+          # :reek:LongParameterList -- Constraint by Elasticsearch's design
+          def initialize(name, field: nil, script: nil, size: nil, order: nil, missing: nil)
             if (field.present? && script.present?) || (field.blank? && script.blank?)
               raise ArgumentError, "Either 'field' or 'script' must be provided"
             end
@@ -42,13 +48,16 @@ module JayAPI
             @script = script
             @size = size
             @order = order
+            @missing = missing
           end
+
+          # rubocop:enable Metrics/ParameterLists
 
           # @return [self] A copy of the receiver.
           def clone
-            self.class.new(name, field: field, script: script, size: size, order: order&.deep_dup).tap do |copy|
-              copy.aggregations = aggregations.clone
-            end
+            self.class.new(
+              name, field: field, script: script, size: size, order: order&.deep_dup, missing: missing
+            ).tap { |copy| copy.aggregations = aggregations.clone }
           end
 
           # @return [Hash] The Hash representation of the +Aggregation+.
@@ -60,7 +69,8 @@ module JayAPI
                   field: field,
                   size: size,
                   script: script&.to_h,
-                  order: order
+                  order: order,
+                  missing: missing
                 }.compact
               }
             end